+++ src/components/BarChart/components/AnnotationLine/AnnotationLine.tsx
@@ -1,4 +1,5 @@
 import React from 'react';
+import type {Direction} from 'types';
 
 import {classNames, clamp} from '../../../../utilities';
 import type {Annotation} from '../../types';
@@ -7,39 +8,54 @@ import styles from './AnnotationLine.scss';
 
 const MEDIAN_OFFSET = 0.5;
 
-export interface AnnotationLineProps extends Omit<Annotation, 'dataIndex'> {
-  xPosition: number;
-  barWidth: number;
-  drawableHeight: number;
+export interface AnnotationLineProps
+  extends Omit<Annotation, 'dataPointIndex' | 'dataSeriesIndex'> {
+  position: number;
+  barSize: number;
+  drawableSize: number;
   shouldAnimate?: boolean;
+  direction?: Direction;
 }
 
 export function AnnotationLine({
-  xPosition,
-  barWidth,
-  drawableHeight,
-  width: annotationWidth,
-  shouldAnimate = false,
+  barSize,
   color,
-  xOffset = MEDIAN_OFFSET,
+  direction = 'vertical',
+  drawableSize,
+  offset = MEDIAN_OFFSET,
+  position,
+  shouldAnimate = false,
+  width: annotationWidth,
 }: AnnotationLineProps) {
   const halfAnnotationWidth = annotationWidth / 2;
-  const offset = barWidth * xOffset;
+  const centerOffset = barSize * offset;
   const xValueClamped = clamp({
-    amount: xPosition + offset,
-    min: xPosition + halfAnnotationWidth,
-    max: xPosition + barWidth - halfAnnotationWidth,
+    amount: position + centerOffset,
+    min: position + halfAnnotationWidth,
+    max: position + barSize - halfAnnotationWidth,
   });
+
+  const xy =
+    direction === 'vertical'
+      ? {
+          x1: xValueClamped,
+          x2: xValueClamped,
+          y1: drawableSize,
+          y2: 0,
+        }
+      : {
+          y1: xValueClamped,
+          y2: xValueClamped,
+          x1: drawableSize,
+          x2: 0,
+        };
 
   return (
     <line
       className={classNames(shouldAnimate && styles.AnimatedLine)}
       stroke={color}
       strokeWidth={annotationWidth}
-      x1={xValueClamped}
-      x2={xValueClamped}
-      y1={drawableHeight}
-      y2={0}
+      {...xy}
     />
   );
 }