+++ src/components/BarChart/types.ts
@@ -1,18 +1,26 @@
-import type {Color} from 'types';
-
-export interface BarChartData {
-  barColor?: Color;
-  label: string;
-  rawValue: number;
-}
+import type {TooltipData} from 'components/TooltipContent';
+import type {LabelFormatter} from 'types';
 
 export interface RenderTooltipContentData {
-  label: string;
-  value: number;
+  data: TooltipData[];
+  title?: string;
   annotation?: Annotation;
 }
+
+export interface XAxisOptions {
+  labelFormatter?: LabelFormatter;
+  hide?: boolean;
+  wrapLabels?: boolean;
+}
+
+export interface YAxisOptions {
+  labelFormatter: LabelFormatter;
+  integersOnly: boolean;
+}
+
 export interface Annotation {
-  dataIndex: number;
+  dataSeriesIndex: number;
+  dataPointIndex: number;
   width: number;
   color: string;
   tooltipData?: {
@@ -20,7 +28,7 @@ export interface Annotation {
     value: string;
   };
   ariaLabel?: string;
-  xOffset?: number;
+  offset?: number;
 }
 
 export interface AnnotationLookupTable {