+++ src/components/MultiSeriesBarChart/index.ts
@@ -1,2 +1,0 @@
-export {MultiSeriesBarChart} from './MultiSeriesBarChart';
-export type {MultiSeriesBarChartProps} from './MultiSeriesBarChart';