+++ src/components/StackedAreaChart/tests/Chart.test.tsx
@@ -14,7 +14,7 @@ import {
 } from '../../../components/TooltipWrapper';
 import {mountWithProvider, triggerSVGMouseMove} from '../../../test-utilities';
 import {StackedAreas} from '../components';
-import {Chart} from '../Chart';
+import {Chart, Props} from '../Chart';
 
 jest.mock('../../../utilities', () => {
   return {
@@ -36,31 +36,31 @@ describe('<Chart />', () => {
     jest.useRealTimers();
   });
 
-  const mockProps = {
-    series: [
+  const mockProps: Props = {
+    data: [
       {
         name: 'Asia',
         data: [
-          {label: '1', rawValue: 502},
-          {label: '2', rawValue: 1000},
-          {label: '3', rawValue: 2000},
-          {label: '4', rawValue: 1000},
-          {label: '5', rawValue: 100},
-          {label: '6', rawValue: 1000},
-          {label: '7', rawValue: 5000},
+          {key: '1', value: 502},
+          {key: '2', value: 1000},
+          {key: '3', value: 2000},
+          {key: '4', value: 1000},
+          {key: '5', value: 100},
+          {key: '6', value: 1000},
+          {key: '7', value: 5000},
         ],
         color: 'purple',
       },
       {
         name: 'Africa',
         data: [
-          {label: '1', rawValue: 106},
-          {label: '2', rawValue: 107},
-          {label: '3', rawValue: 111},
-          {label: '4', rawValue: 133},
-          {label: '5', rawValue: 100},
-          {label: '6', rawValue: 767},
-          {label: '7', rawValue: 1766},
+          {key: '1', value: 106},
+          {key: '2', value: 107},
+          {key: '3', value: 111},
+          {key: '4', value: 133},
+          {key: '5', value: 100},
+          {key: '6', value: 767},
+          {key: '7', value: 1766},
         ],
         color: 'teal',
       },
@@ -71,7 +71,6 @@ describe('<Chart />', () => {
     formatXAxisLabel: (val: string) => val,
     formatYAxisLabel: (val: number) => val.toString(),
     renderTooltipContent: jest.fn(() => <p>Mock Tooltip Content</p>),
-    colors: ['purple', 'teal'],
   };
 
   it('renders an SVG', () => {
@@ -214,7 +213,7 @@ describe('<Chart />', () => {
     const chart = mount(<Chart {...mockProps} />);
 
     expect(chart).toContainReactComponent(VisuallyHiddenRows, {
-      series: mockProps.series,
+      data: mockProps.data,
       xAxisLabels: mockProps.xAxisOptions.labels,
       formatYAxisLabel: mockProps.formatYAxisLabel,
     });